feat: extend isWholeNumber to float and double
