+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/internal/constants/deprecations.kt
@@ -1,0 +1,3 @@
+package xyz.lbres.kotlinutils.internal.constants
+
+internal const val basePackage = "xyz.lbres.kotlinutils"
