+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/number/BigDecimalExt.kt
@@ -1,5 +1,6 @@
 package xyz.lbres.kotlinutils.number
 
+import xyz.lbres.kotlinutils.utils.succeeds
 import xyz.lbres.kotlinutils.utils.tryOrDefault
 import java.math.BigDecimal
 
@@ -16,6 +17,13 @@ fun BigDecimal.isZero(): Boolean = tryOrDefault(false) { toBigIntegerExact().isZ
  * @return [Boolean]: true if value is less than zero, false otherwise
  */
 fun BigDecimal.isNegative(): Boolean = this < BigDecimal.ZERO
+
+/**
+ * Unary check to determine if value is a whole number
+ *
+ * @return [Boolean]: true if value is a whole number, false otherwise
+ */
+fun BigDecimal.isWholeNumber(): Boolean = succeeds { toBigIntegerExact() }
 
 /**
  * Unary check to determine if value is null or zero
