+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/number/DoubleExt.kt
@@ -26,6 +26,13 @@ fun Double.isZero(): Boolean = abs(this) == 0.0
 fun Double.isNegative(): Boolean = this < -0.0
 
 /**
+ * Unary check to determine if value is a whole number
+ *
+ * @return [Boolean]: true if value is a whole number, false otherwise
+ */
+fun Double.isWholeNumber(): Boolean = rem(1).isZero()
+
+/**
  * Unary check to determine if value is null or zero
  *
  * @return [Boolean]: true if the value is null or zero, or false otherwise
