+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/number/bigdecimal/BigDecimalExtTest.kt
@@ -1,6 +1,5 @@
 package xyz.lbres.kotlinutils.number.bigdecimal
 
-import xyz.lbres.kotlinutils.testutils.checkTrueFalse
 import java.math.BigDecimal
 import java.math.BigInteger
 import java.math.RoundingMode
@@ -8,25 +7,6 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class BigDecimalExtTest {
-    @Test
-    fun testIsWholeNumber() {
-        val trueValues = listOf(
-            BigDecimal.ZERO,
-            BigDecimal("100"),
-            BigDecimal("-100"),
-            BigDecimal("-1234560000000000000999"),
-            BigDecimal("123456000000000.0000"),
-            BigDecimal("7.00000000000000000000000000000"),
-        )
-        val falseValues = listOf(
-            BigDecimal("0.00000000000000000001"),
-            BigDecimal("-1.01"),
-            BigDecimal("123456000000000000099.9"),
-            BigDecimal("99999999999999999999999.9"),
-        )
-        checkTrueFalse(trueValues, falseValues, { "$it.isWholeNumber()" }, BigDecimal::isWholeNumber)
-    }
-
     @Test
     fun testRoundToBigInteger() {
         // whole
